chore(metrics): Rename rds_replication_slot_disk_usage_average to rds_replication_slot_disk_usage_bytes

+++ internal/app/rds/rds.go
@@ -214,9 +214,9 @@ func (r *RDSFetcher) computeInstanceMetrics(dbInstance aws_rds_types.DBInstance,
 
 	iops, storageThroughput := getStorageMetrics(*dbInstance.StorageType, int64(dbInstance.AllocatedStorage), iops, throughput)
 
-	var maxAllocatedStorage int32 = 0
+	var maxAllocatedStorage int64 = 0
 	if dbInstance.MaxAllocatedStorage != nil {
-		maxAllocatedStorage = *dbInstance.MaxAllocatedStorage
+		maxAllocatedStorage = int64(*dbInstance.MaxAllocatedStorage)
 	}
 
 	pendingModifiedValues := false
@@ -271,7 +271,7 @@ func (r *RDSFetcher) computeInstanceMetrics(dbInstance aws_rds_types.DBInstance,
 		Engine:                     *dbInstance.Engine,
 		EngineVersion:              *dbInstance.EngineVersion,
 		LogFilesSize:               logFilesSize,
-		MaxAllocatedStorage:        converter.GigaBytesToBytes(int64(maxAllocatedStorage)),
+		MaxAllocatedStorage:        converter.GigaBytesToBytes(maxAllocatedStorage),
 		MaxIops:                    iops,
 		MultiAZ:                    dbInstance.MultiAZ,
 		PendingMaintenanceAction:   pendingMaintenanceAction,
